[CUB-4622] Move index to src
