+++ package.json
@@ -8,7 +8,7 @@
   },
   "license": "Apache-2.0",
   "author": "labforward GmbH",
-  "main": "index.ts",
+  "main": "dist/index.js",
   "scripts": {
     "build": "del-cli dist && yarn build:schema && tsc -p tsconfig.build.json && yarn copy-dts",
     "build:schema": "ts-node src/compile.ts",
